Passing additional surface variables into the atmospheric model #28 (#29)
* ADDED SURFACE VARIABLES REQUIRED BY NCEP TKE-EDMF

The shflx, lhflx, wind, thv_atm, thv_surf variables are passed
from the surface coupler to the atmospheric model as required by
the NCEP TKE-EDMF scheme.

* CONDITIONAL COMPILATION FLAG FOR AM4 COMPATIBILITY

For compatibility with AM4, a ifdef-statement with a conditional
compilation flag "do_am5phys" is added in the call to
physics_driver_down to prevent the additional surface variables
from being passed into the old AM4 physics package. This flag
should be defined if AM5 is compiled with NCEP TKE-based EDMF.

* REMOVED NAME TAGS FROM THE COMMENTS

* REMOVING THE CONDITIONAL COMPILATION FLAG

The added variables are now passed as optional arguments into
physics_driver_down. This requires consistent changes in the
atmos_phys package.

* ADDED IF-STATEMENTS FOR CHECKSUM

Two if-associated statements are added before the checksums
of bnd_type%shflx and bnd_type%lhflx to pass the null model
build.

* Changed the name for EDMF in the comments

The full name for the NCEP TKE-based EDMF replaces the informal
short name (HanEDMF) in the comments only. No other code changes.

* Modified the comments for new fields

The term 'surface' in the comments for the added fields is
clarified as the lowest atmos level as Sergey suggested.

---------

Co-authored-by: Zhihong Tan <Zhihong.Tan@gaea52.ncrc.gov>